requirements.txt
django==4.2
psycopg2==2.9.10
python-dotenv==1.1.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-admin-rangefilter==0.13.2
django-cors-headers==4.7.0
pyjwt==2.9.0
drf-yasg==1.21.10
huey==2.5.3
redis==6.2.0
